apps/preparation/pw-l1c/src/main/java/eu/csgroup/coprs/ps2/pw/l1c/service/L1cPWInputManagementService.java
/*
 * Copyright 2023 CS Group
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package eu.csgroup.coprs.ps2.pw.l1c.service;

import eu.csgroup.coprs.ps2.core.common.model.l1.L1ExecutionInput;
import eu.csgroup.coprs.ps2.core.common.model.processing.ProcessingMessage;
import eu.csgroup.coprs.ps2.core.common.model.trace.TaskReport;
import eu.csgroup.coprs.ps2.core.common.model.trace.task.ReportTask;
import eu.csgroup.coprs.ps2.core.common.settings.MessageParameters;
import eu.csgroup.coprs.ps2.core.common.utils.ProcessingMessageUtils;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;

@Slf4j
@Service
public class L1cPWInputManagementService {

    public L1ExecutionInput manageInput(ProcessingMessage processingMessage) {

        log.info("Received message {}", processingMessage);

        TaskReport taskReport = new TaskReport()
                .setTaskName(ReportTask.PRODUCTION_TRIGGER.getName())
                .setSatellite(processingMessage.getMissionId() + processingMessage.getSatelliteId());

        taskReport.begin("Received L1 Execution message");

        L1ExecutionInput executionInput;

        try {
            executionInput = extract(processingMessage);
            taskReport.end("Successfully extracted processing input");
        } catch (Exception e) {
            taskReport.error("Error managing input for message " + processingMessage);
            throw e;
        }

        return executionInput;
    }

    private L1ExecutionInput extract(ProcessingMessage processingMessage) {
        log.info("Extracting input from message");
        final L1ExecutionInput l1ExecutionInput = ProcessingMessageUtils.getAdditionalField(
                processingMessage,
                MessageParameters.EXECUTION_INPUT_FIELD,
                L1ExecutionInput.class
        );
        log.info("Finished extracting input from message");
        return l1ExecutionInput;
    }

}
